requirements.txt
# Vietnamese Voice Chat - Minimal Requirements
# Size: ~268MB (down from 1.2GB)

# Core web framework
flask>=2.0.0
flask-cors>=3.0.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=0.19.0

# Audio processing
soundfile>=0.12.0
numpy>=1.21.0
pydub>=0.25.0
librosa>=0.9.0

# Text-to-Speech
gtts>=2.3.0

# Utilities
requests>=2.25.0
urllib3>=1.26.0 